reorganize tests, index file and severeal smaller files with testing the same functionality
