+++ test/index.test.js
@@ -1,5 +1,9 @@
-const routes = require('./routes.test.js');
-const backend = require('./backend_test.js');
+const routes = require('./lib/route_test.js');
+const server_test = require('./lib/server_test.js');
+const wiki_test = require('./lib/wiki_test.js');
+const nomad_test = require('./lib/nomad_test.js');
 
 routes();
-backend();
+wiki_test();
+server_test();
+nomad_test();
