+++ test/lib/route_test.js
@@ -1,4 +1,4 @@
-const server = require('../src/server.js');
+const server = require('../../src/server.js');
 const tape = require('tape');
 
 function test () {
